+++ changelog.md
@@ -5,6 +5,7 @@ Parsing and sanitization improvements are always ongoing and non-breaking
 ## 1.2
 
 - Added `nearest(lat, lon)` and `sends_reports` to `Station`
+- Added `nearest(lat, lon, n)` to `station`
 - Added lazy loading handling
 - Moved `Station`, `uses_na_format`, `valid_station` to `avwx.station`